Implements secure djangoauth login (closes #184)
To ensure security, the djangoauth plugin now requires the "sessionid"
cookie to be set on all requests from the client that require
translation to and from a user's id.  It will then use this cookie to
request login information from django directly over an HTTP GET request.

The exception to this is when an "invite" action is requested: in this
case, invited player_id's will be translated (and created) using the old
method. (The owner_id is secured, though.)