Implements standalone django plugin tests (closes #189)
All django-dependent cardstories plugins (djangoauth and mail, at the
moment) can now run successfully without requiring a django server to be
running.

This was implemented using django's built-in unittest API, and as a
consequence, disabling CSRF mode in settings.py is no longer required.

Another consequence of this implementation is that the website django
application itself now runs on a safer namespace ("website.*"), without
which it was impossible to run the test framework outside django itself.