Inplements Facebook login and registration (closes #149)
This implements Facebook login and registration through a Facebook login
button in the welcome page, in addition to the login and registration
forms.  No external libraries were used, and the OAuth server-side
authentication flow was chosen, as well as the Graph API for retrieval
of user information.  See the following:

https://developers.facebook.com/docs/authentication/
https://developers.facebook.com/docs/reference/api/

Behavior: when clicking on the login button, the user will be redirected
to Facebook and the Cardstories app will ask for basic and email
permissions.  If authorized, the user will be redirected to Cardstories,
where (s)he will be automatically registered with their email and name
(just as with the regular registration form).  In addition, though,
their Facebook Id is also stored for future reference.

Note to developers: the following must be met in order to be able to
work with this.

1) a Facebook app id and app secret must be specified in settings.py.

2) An entry on /etc/hosts must be created with the following (or
whatever domain name was chosen when running syncdb interactively, or
through the admin interface):

'127.0.0.1 local.cardstories.org'

3) The local development server must be run on port 80 (thus, as root).