Implements support for OWA (refs #167)
This implements support for Open Web Analytics both in the website and
in the Cardstories client, in the latter case through a simple
publish/subscribe architecture of stage changes (primarily so that
the client plugin can remain agnostic, and have OWA-specific code live
in a different plugin).

The default behavior is to have OWA itself disabled.  See new
instructions in README.txt for how to enable it.