Implements progress mark animation (closes #258)
This implements an animate_progress_bar() function in the cardstories
client that should be called at the end of a given state, so that the
progress mark at the top will animate towards the next position.

It also implements this function for the transition from the first to
the second state.