Integrates jqDock morph animation (closes #273)
This integrates the jqDock animation (refs #246) with the info box
entry scaling (refs #249), in such a way that the second happens after
the first.  Unfortunately, jqDock does not provide a hook for when the
"expand" animation finishes, so a simple timeout had to be used.

A bug was also fixed on $.cardstories.animate_scale where it would not
handle undefined callbacks safely.