Implements create_pick_card modal box behavior (closes #250)
Implements the info/modal box behavior in such a way that while the user
does not click on "OK", the dock is essentially frozen (through a modal
overlay) and no cards can be selected.

When "OK" is clicked, the info box is shrunk away and the modal overlay
is disabled.