Fixes matching "/resource" on the dev server (refs #289)
The "/resource" url on the website dev server must match the one in use
by the cardstories client; it is sensitive to a missing "/" (which was
the problem).

Also reverts the test djangoauth host URL, as it more suited to a
complete development environment that will also works with the CS
Facebook app (when it goes live).