Replace JS code for image buttons with CSS.
The wrapper that is holding the images should have
a class of `cardstories_img_button`. It should include
at least two images, one with `cardstories_btn_on` class
(the usual state) and one with `cardstories_btn_over` class,
for the hover state.

If the button can be in a disabled state, it should include
a third image with a class of `cardstories_btn_off`,
which will be visible if the wrapper div has a class
of `cardstories_button_disabled`.

Because we rarely use separate images for the active state,
they need to be implemented on a case to case basis.
The "invite friend" button slots may serve as an
example of this.

Closes #329.