Refactors animate_progress_bar (closes #312 and #307)
This refactors animate_progress_bar() into a less cumbersome (and
therefore more useful) abstraction.  All it needs now is the destination
step and the current element.

To do so, all the progress bar mark positions are now predefined in the
stylesheet.  There is no longer a need to define each mark's position
per state.

The progress bar animation from step 3 to step 4 was also implemented:
it happens as soon as the first invited player joins the game.

Finally, a bug was fixed where the progress bar would be displayed more
than once after every change in game state.