Implements author's results box (closes #375 and #376)
Displays the author's results, according to how the game was won or
lost, and enables the "play again" button.

Also fixes a bug where if the players won a game that was "too hard",
their names would be colored red on the board.

Finally, adds back Loic's play_again unit test, with minor
modifications.