Use safer positioning strategy (closes #383)
Refactors most of the client code to use jQuery's position() instead of
css('left') or css('top').  Although position() requires elements to be
visible in order to work, this is a safer cross-browser way of doing
things.  In particular, Firefox 3.6 and IE 8 seem to behave much more
coherently with it.