Fixes animation sync issues (closes #412)
This extends the current deferred architecture so that for the invite
and vote game states, the poll is only called when the deferreds are
successfully resolved.  This gives us:

1) A mechanism to signal IF we want the poll to run (sometimes not
necessary, such as in the invitation_participate state, which calls
send_game anyway).

2) A method to signal WHEN we want the poll to start, so that we can
delay it until critical initial animations are over.