Implement transition from pick_wait to vote_voter.
The transition is impossible to implement in a satisfying
way when some of the players don't pick a card in time.
Those players don't appear in the game.players array anymore
and so the position of player slots is changed in a way
that cannot be animated well.

For the time being, those players are only faded out
during the transition. In the future it would be good
to modify the server code so that it doesn't completely
remove viewers/anonymuos players from the game.players
array, but only marks them as such.

A ?skin=pick_wait_to_vote state that simulates the transitions
when one of the players doesn't pick a card in time was added,
so that we don't forget this.

Closes #417.