Fixes card preservation in standalone mode (closes #422)
This is a situation where it is better to rely on javascript's "=="
comparison than on strict "===".  When using numeric player_id's (such
as when testing using ?player_id=1, or ?player_id=2), these tests would
fail where they should succeed, with the symptom that when moving from
pick to pick_wait, and from pick_wait to vote_voter, the selected card
would not be preserved.