Fixes game_id not found on complete (closes #435)
This fixes the bug where a "game_id not found" alert would be issued to
a player because simultaneous polls were being queued by the browser
(more details in the bug report).

To do this, the poll logic was changed so that there can never be more
than one happening simultaneously.  If necessary, the XHRs can be
aborted by poll_discard() (notably, to cancel the lobby poll when
starting a story).

Tests were refactored accordingly.

Also, calls to console were removed: they were breaking IE (it's no use
verifying if console exists, IE will still barf if you mention it).