Implements chat subsystem (refs #496)
This pulls together all the previous pieces, creates some new ones, and
glues all of them together to get the first version of a chat working.
At this point, there are no avatars, notifications, or list of online
players, but everybody should be able to chat with each other.

Of note:

* An actual plugin mini-infrastructure was created.  Plugins have a few
  things they need to do to let themselves be known to cardstories
  itself: see the chat plugin.

* There were a couple of bugs in chat.py, they were fixed.

* There was lots of refactoring in some critical areas, including the
  fact that now, a poll needs to be running essentially all the time.

* Instead of passing "modified" around, I decided to save it in root as
  metadata.  Otherwise, it would not be possible to asynchronously know
  when the latest modification had taken place (given that on some
  situations the screen is not redrawn when the poll returns).