Show a warning for non-existant game_id (closes #499).
Instead of erroring out with an Internal Server Error
when passes a non-existant game_id, alert the user
with an explanation.

I am sure there is a better way to write the test,
but being a twisted noob, I gave up trying to come
with a more idiomatic solution after a couple of
unsuccessful attempts.