Implement the audio plugin (closes #514).
This commit implements the cardstories_audio plugin
with a play(sound_id) method. It uses SoundManager2
under the hood.

Sounds are defined as metadata on the root element,
structured as a map (JS object) of sound_ids to urls.
I chose to not depend on the jquery metadata plugin,
but use plain HTML5 data attributes and the jquery .data
method.

Due to the unfortunate way that SoundManager2 gets
initialized, the path to the SWF files and a couple
of other SoundManager propeties need to be set very
early in the loading process and therefore cannot be
read from HTML metadata, but need to be hardcoded
in the script fragment of the loading file.
Examples can be seen in static/index.html,
static/audio_example.html, and static/test/audio.html.

Preloading hasn't been considered yet, we could implement
it if it turns out we will need it in practice.