Implements new game notification (refs #509, #512, #513)
This implements a new game notification in the chat, seen when a new
game is created.  To do so, chat messages now have a "type" attribute:
either "chat", or "notification", each with specific attributes.

Important: the chat plugin must come AFTER the auth plugin (either auth
or djangoauth), in order for proper player id translation to occur.  The
README was modified accordingly.

Of note:

* service.game_notify was modified to include the "sentence" when
calling game_init(), otherwise we'd have to make another call to the
database to find out that piece of information about the game.

* The chat plugin now registers itself to receive game-specific
notifications, such as, in this case, "init".  This was modelled after
the mail plugin.

* Substituted all <font/> bold tags with <strong/>, as it is proper
semantic syntax.

* Fixed a bug in chat.py where it was sending out references of
messages, instead of copies.  This had the effect that if a second
plugin post-processed these messages, they would be altered in the
message database.

* The client-side plugin was updated to reflect the above changes.