Play sound on new game notification (closes #515).
To avoid playing the ringing sound for old announcements
on page refresh, we check that the message happened
more than 5 seconds after page initialization before
playing the sound. This is a bit hackish and we might
want to introduce more metadata into the data structure
passed to the chat.js plugin.