Fix hanging tests in IE8 (closes #505).
The problem was that IE8 was parsing "javascript://" url on
a form element as the end of the form tag.
The submit button was therefore seen by IE8 as not being part
of the form and therefore form submission didn't work properly.

To fix this, I replaced "javascript://" with "javascript:void(0)".
I am not a big fan of either, though. I wonder if we could do
without all of these weird javascript-type urls.