Scroll chat div to bottom on window load (closes #526).
This is the second try to fix #526.

The idea is that since chat messages may be loaded before
the window is fully ready (all the images have been loaded, etc)
the dimensions of some elements are not known yet. From experience
I know this to be true at least on chrome.

This commit adds a handler to the window load event to scroll
the chat to bottom. We will need to test and see if this works
in practice since I cannot reproduce the issue on my local machine.