Don't insert preloaded images into DOM (refs #528).
Some browsers don't cache images if they are dynamically
generated, but not inserted into the DOM.

I forgot which browser it was exactly that needed this workaround,
but if I remember correctly it was one version of Firefox.
Firefox 7 seems to be able to cache images fine without
inserting them into the DOM.

Not inserting preloaded images into the DOM seems
to help with the memory usage a bit.