Fix buggy animation on play again (refs #447).
There are multiple divs with the same name, when fading
them out collectively, the callback gets called multiple
times (once for each element).

Be more selective when fading out results so that the callback
only gets called once.

Removed a large part of test that was testing behaviour that
we don't really want/need since the new interface.
(More work on this will be donw for #95).