Show users' names instead of emails (closes #464).
I had to modify the mail plugin a bit, because it was
relying on getting emails through auth.resolve.

I also made a couple of tiny css tweaks to make names look
better because they will now usually be displayed in
two lines and not only one line as was happening up
until now using emails.