Allow players to have same names (closes #619).
When two or more players share the same name,
I've identified three problems with the client:

- The owner was determined by comparing game.owner_id
  with the player_id slot of each item in the game.players
  array. When one of the player had the same name as the owner,
  this led to problems.

- Similarily, the position of current player in the
  game.players array was determined by comparing the player_id
  parameter with the player_id slot of each item in the
  game.players array.

- During transition from pick to vote, and from vote to
  complete states, players may be removed from the game
  (if they don't pick/vote in time). In order to show a
  fade effect on players that have been removed, the client
  needs to identify the players that have been removed
  during the transition. Since it was doing this by comparing
  player_ids from the previous game.players array with the
  new game.players array, it was bound to errors in case of
  any player sharing his name with a removed player.

In order to remedy the situation, each player needs to be uniquely
identified and the owner position in the game.players array needs
to be known.

In order to identify the owner I could have added the unique
numerical ID of the owner to the game info, but I instead decided
to add owner_index (which represents the index position of the owner in
the game.players array - and is usually (always?) zero), because
chosing this option I could simplify code on the client.

For the same reason, I added player_index, representing the
position of the current player in the game.players array.

In order to uniquely identify each player, I decided to simply
add the unique numerical ID of each player to the game.players
array elements. There is a high chance we will need unique player
IDs for future updates anyway.

I added the numerical player ID as the last item in the game.players
array elements, because in this way I didn't have to change a lot of
code, but I am really not happy with this and I would very much like
to clean up the elements of the player elements in the game.players
array, because even without the numerical player ID addition, the
ordering of the data is rather strange.

I would like the order to look like this, which would resemble
the ordering of the game states/events:

[player_id (unique), player's name, player's cards, pick, vote, win 'y/n']

While this would be a simple mechanical work, it would require a
relative large amount of time, because the game.players info is used
in some many places in the code and in the tests.

I would also like to add that having multiple players share the same
name only works when using djangoauth, it doesn't work with the plain
auth plugin.