Fix chat scrolling (closes #526).
Hopefully this will finally fix the annoying scrolling
issue. This time I have been lucky enough to successfully
reproduce the issue on my local computer (at least I hope
I reproduced the same issue).

The chat didn't scroll when moving from the lobby to the
create new game state. Previously, function scroll_to_bottom()
of the chat plugin wasa called for every received chat message.
When the state function was called without any messages (that's
what happens when going from the lobby to the new game state),
the scroll_to_bottom() was not called at all, which was the
source of the problem.

It was fixed by simply moving the scroll_to_bottom call out
of the loop below it. This also has a benefit of only calling
scroll_to_bottom once per a received batch of chat messages,
calling it once for each of the messages was completely redundant.