Return the poll immediately if game is destroyed (closes #611).
This fixes a race condition where a poll request might
not reach the server before the game has already been
destroyed (because it has been completed).

When a poll request reached the service and the requested
game wasn't loaded in memory, the service raised an error.
That was changed so that it now returns the poll immediately
and doesn't raise any errors.