Check current state before picking/voting (closes #606).
Previously, the game didn't check the game state before
executing pick and vote requests, which allowed for rare,
but nasty race conditions where a player would vote after
the results have already been calculated, which led to
buggy behaviour of the client in the 'complete' phase.

Similar race conditions were possible when picking a card
when the game has already moved to the 'vote' state.

The game now checks in a transaction to make sure it
is currently in 'invitation' (for pick) or 'vote'
(for vote requests).

If current state is not the expected state, the game
raises a UserWarning with an explanation.