Don't alert empty messages (closes #654).
If an ajax error occurs and the error text is
not present, the alert box is not displayed,
and the ajax request is retried automatically
in the background (while the error is also
logged to the console, if window.console is
present).

It remains to be seen if this is a good strategy
with errors that happen in real life.