Make most ajax request asynchronous (closes #555).
This commit makes most ajax requests issued by
the client asynchronous.

The common defaults have been incorporated into
$.cardstories.ajax function so that they don't
need to be repeated in every single ajax call.

In some cases where the request is initiated by the
user, the requests are kept synchronous.
This is so because it is easier to make sure no race
conditions occur, while it also informs user that something
is happening in the background by blocking the UI.
If the request that sends the countdown duration didn't
block, for instance, and the user was impatient due
to a slow connection, he could keep changing the dropdown
and firing ajax requests. If they are syncrhounous, we
can be sure that they will hit the server in order and
the last selected value will be chosen. If they were async,
we could say nothing about the ordering and the selected
value on the server would be arbitrary.

In cases where the user initiates sync requests himself,
there is also no worry that he will be surprised by chat
input blocking, since he knows something is happening in
the background. The same cannot be said for requests such
as polling or queued requests, that fire after animation
finishes. The user can't expect those, and cannot understand
why the chat input would block while he is happily typing
messages.