Only play pop/ring sounds once per message batch (closes #681).
It turns out I was not right about sounds being played
multiple times on refresh, because the initialized timestamp
was supposed to take care of that in most cases.

Still, this commit was worth it also for the cleanup and
adding the pop sound to the index.html structure
(spent some time wondering why the pop sound doesn't
want to play when playing without django :) ).