Closes #720 - Destorying games should not delete data about players
Previously, when a game was destroyed (timer expiration), the
player2game data was deleted from the database. This would prevent
the bots from using data from those games, since it wouldn't be
possible to know which card the author had chosen in relation to
the game sentence.

On the other side, game expiration (currently set to one month on
the live server, so in practice never applied to any game) need to
be active - otherwise bots try to join all open games from the
database (>500), and in any case the open games are cluttering the
chat log upon server restart.