Display notification when accessing canceled game (closes #711).
I added a separate HTML section just for notifications.
I was thinking maybe we will need more of these, maybe
for the user warnings, although I am not sure if they
will all fit into this pattern... There will probably
be some messages/notifications that should be less
disruptive to the game than showing a popup dialog.

If we decide this was a useful idea and will keep
the cardstories_notifications section, it might make
sense to take the logout button out of the progress bar.
Right now, the logout button is only shown when the
`display_progress_bar` function is called.
For notifications we don't call this function (there
is no game and therefore no progress), so the logout
button is not seen.
I let this slip for now as it is a minor issue
and I am not yet convinced we will be using the
cardstories_notifications section much.