Alert a meaningful error message when voting too late (closes #722).
This is a veeery hackish workaround to display a more meaningful
messages when a user votes too late when the game is already in
the complete state. Previously it alerted "game doesn't exist"
which was confusing and wrong.

This hack is temporary and in order to fix this problem properly
(and generally), I created a separate (larger) ticket: #732.