Return error codes instead of strings from service (closes #732).
Previously the service raised UserWarnings with a string message
on errors, that were handled by the service by simply alerting
the message to the user.

This was changed so that errors that are supposed to be handled by
the client return an error code, based on which the client can perform
its action.

The errors that we are raising in the service were split into two
cases:

  - error conditions that we expect to happen sometimes (ie. as a
    consequence of unavoidable race conditions) are raised as
    instances of CardstoriesWarning, with a proper error code and
    an ooptional dictionary of additional data. We expect the client
    to be prepared for these errors/warnings and handle tham in a
    graceful manner.

- error conditions that we never expect to happen under normal
    operation of the service. The errors are instances of CardstoriesException,
    but are handled the same as any other unexpected exception.
    In case they are raised, the client will alert them, but not
    try to do anything else.

The $.cardstories.error function was replaced by $.cardstories.panic
and $.cardstories.show_warning functions.

The $.cardstories.panic function is used in situations where the client
doesn't know of a better way to handle the error, so it just alerts it to
the user (similar to what the $.cardsotries.error function did before).

The $.cardstories.show_warning function is used for displaying warnings
as an in-game modal, and accepts a callback that that will be invoked
after the user closes the modal by clicking the button.

When a warning is displayed, the top bar with game progress and the logout
button are not visible. This is a consequence of how the HTML is currently
structured. Ideally, the progress bar and the logout button would remain
visible while showing the warning, but this would require a major rewrite
of the HTML and associated display logic.

The four warnings that are currently handled by the $.show_warning function
were added to the ?skin page. The skin.js seems to have been significantly
changed from looking at the diffs, but apart from the added logic at the end
of the file that handles the new "error type" skins, it was just that I let
emacs re-indent the file automatically (because indentation was very inconsistent)
and fixed some of the errors JSLint complained about while chasing bugs.

To make some operations easier, I also changed the $.cardstories.set_active
parameter order, and made $.cardstories.reload_link, $.cardstories.reload,
and $.cardstories.init_board functions not take player_id as a parameter
anymore, since that parameter wasn't used by any of them anymore anyway.

I was having significant problems with JS tests. Therefore I updated QUnit,
did some JSLint-ing, and made the tests setup function re-initialize the
$.cardstories object from an intact copy before every test is run. This way,
we don't have to remember to restore functions that we stub during tests.
Forgetting to unstub a function has proved to be one of the most frequent
reasons for weird flip-flopping test failures.

After I made these changes, the tests pass everywhere, except IE8, where they
sometimes pass, but more often randomly fail with some of the weirdest
error messages I have seen so far. Since the errors looked a bit similar
to the problems I once had with "javascript://" type links, I changed
all occurances of "javascript://" with "javascript:void(0)" in the HTML,
but this didn't seem to fix the problem. So I gave up on IE8 and will
create a seperate issue for this.

As a final note, It would probably be good to skin the warning modals
a bit differently from the other modal dialogs, perhaps by adding a
warning sign of some sort, or changing their color.