Implement 'multigame' type poll (closes #756).
This also implements a multigame state call, to be able
to get state of multiple games in one call to the service.

The idea is that once the multigame poll returns,
you should be able to pull the states of all games that
are "opened" in the tabs and update the tabs accordingly.

It might be a good idea to pull this out into a separate plugin.