Implements XP data model (see #767)
This implements the player experience data model through the current
authentication plugins.  In other words, the auth backend will be used
to store all player info related to score, leveling up, and earned
cards.  (This data is planned to be managed by the webservice through
yet-to-be-implemented getter/setter methods.)

Score vs. levels: since one of the requirements is to have a tunable
level formula, the actual level will not be stored in the database -
only the current score (through the "score" field).  However, another
requirement is that if the level-up difficulty is raised, players should
not necessarily earn cards on the next level.  To honor this, the
"levelups" field is used to store how many times the player has leveled
up - although at any given moment it may have the same value, this
field's content does NOT reflect a players actual level (which is always
calculated according to the current formula, planned to be hard-coded in
game.py).

Earned cards: these will be stored on a many-to-one relationship to the
registered users.  As opposed to a VARCHAR field, this makes the
information much easier to manage, specially from the Django admin
interface (a feature that is already implemented by this patch).

Card shuffling: although not readily apparent because no schema changes
are needed, the current card shuffling mechanism (as seen on
CardstoriesGame.createInteraction and .participateInteraction) will need
to be changed in such a way as to limit each players hand to the cards
they have access to.  To do so, the `games.cards` field will be used to
store which cards have already been chosen, as opposed to which cards
are still in the deck.  On each successive joining, a new set of cards
will be built containing the default cards and the ones the player has
earned, minus the ones that have been chosen.  This set will then be
shuffled, a hand extracted accordingly, and the result saved.