Fixes #755 - Implement monitoring of player online status
Implemented as a plugin, which monitors the active polls on the chat
plugin (which is currently always active when the player is logged in).

The activity plugin allows to request a list of online players using a
state request of type 'activity' (can be used later on to implement the
list of online players in the client), or other plugins through the
Observable interface, by sending notification each time a player logs in
or logs out.

Players who drop polls are given time to reastablish one, to avoid false
positives when a poll naturally ends (timeout or notification), or when
a player refreshes the page or switch to a new game.