Fixes #757 - WS plugin for tables
Keeps track of the games online players are playing, facilitates
finding games to join, and helps a group of players to keep playing a
series of games together.

The plugin answers state requests of the type 'table', answering with
the available information on the next game the player should
automatically join. It bases his decisions on the games which are
currently in the 'invitation' stage, and who have players currently
online, as determined by the activity plugin.

* If the player isn't currently in a game, the plugin will try to find an
  available table/game for him to join ('next_game_id') or ask the player
  to create a game ('next_owner_id' set to the player's id).
* Once in the game, upon completion of a game, the webservice will
  decide who should create the next game (each player in turn). The
  information returned contains the player_id of the next owner
  ('next_owner_id'). When another game is created with the
  'previous_game_id' set to the current active game, it becomes the
  active game of the table, and the state includes its id in
  'next_game_id'.

The plugin is pollable, to allow the client to be notified:
* When the players are waiting for the next_owner to create a game
  (polls are then specific to each table).
* When waiting for any table to become available (global poll)