Closes #760 - JS client table plugin
Handles client-side interaction with the WS for tables. It takes care of
setting up a poll on the WS through the plugin interface, for the
current game, and handles the state() updates.

Upon initialisation, if no game is currently selected and if game
creation isn't explicitely requested through the ?create=1 argument, it
requests an available table/game from the WS and redirects the browser
to it if any. (Otherwise, it simply let the client proceed to the
default action, which is to create a new game).

Since the table plugin keeps the plugin up to date on the table
information, the data received through the state() method is stored on
the root, to allow direct access when needed.

The plugin also exposes methods to allow to redirect to the next game
when needed ('load_next_game_when_ready()'), which either redirects
directly, or waits for the information about the next game to arrive if
it isn't there yet.