First implementation of tabs (closes #768).
This commit implements the first version of tabs.

While the server side code feels rather finished and isn't expected to
change much, the client code will need a lot of tweaking until it is ready
for production. Because I wanted to finish this task as quickly as
possible, I didn't pay any attention to styling or even position of the
tabs, I just dropped them to where it was the most simple to put them
(on the right side, under the game board), but that obviusly is not
a good place for tabs. The idea is that the styling/positioning
will be done in a separate task.

Tabs that a user keeps open are stored in a new 'tabs' table on
the webservice. If a player visits a game, for which it doesn't
already keep a tab open, the game is automatically added to the
tabs table. Tabs are deleted from the tabs table through a webservice
endpoint, which is called when the user clicks a close button on a tab.
This is currently the only way tabs get removed, there is no auto
expiration/removal of tabs implemented yet.

Tabs are currently just links pointing to games. Clicking on a tab,
reloads the page. While this isn't optimal from the user experience
point of view, it was by far simplest to implement.

Tabs that require action currently get marked with an asterisk.
The story on the tracker mentions different colors, but since
I didn't want to mess with CSS for this task, I simply made it show
an asterisk. This should be very easy to change and is marked
with a TODO note in the code.

For some of the decisions, we need to hold some state of tabs.
For instance, a complete game should require action when it enters
the complete state, but only until the user hasn't seen the results yet.
Currently the only state that the tab holds is the previous state of the
game (invitation, vote...), and a completed game is implemented to only
require action if previous state was not complete. This isn't ideal, though,
since if multiple changes accross tabs happen in a short period of time,
the complete game will stop announcing it requires action.
A better solution for the complete game would be to remember whether the
user has already seen the results, but since this is more complicated
than the simple solution implemented now, I didn't want to include it in
this task.

The rules that determine when a tab requires an action will probably
still change as we get more feedback and try the tabs out in action,
and shouldn't be hard to modify, since the tab plugin gets to examine
the full state of each game when deciding whether it should consider
the tab as requiring action. The rules are hardcoded inside the JS
tabs plugin, inside the requires_action function.

For the tabs to work, a new table ('tabs') needs to be created in
the database. I added an sql script inside the 'migrations' folder
that can be fed to the sqlite CLI tool to create the table (and the
associated index):

    $ sqlite3 /path/to/cardstories.sqlite < migrations/001_add_tabs_table.sql

While working on this feature, I found errors were hard to debug,
since the service didn't return information detailed enough on
unexpected errors, so I added a stacktrace to the returned data
(previusly it was only the error message).