Fix buggy tab closing behaviour (closes #778).
I copied the behaviour from Chrome - when the current tab is closed,
and there are still other tabs on the right side, it will load the first
tab on the right.
If there are no tabs on the right, but there are tabs on the left side,
it will load the first one to the left.
If it is the last tab, it will open a 'New game' tab.

I didn't prevent from closing the last tab, even if the last tab is 'New game',
because it looks like there are other problems with the 'New game' tabs
that we need to consider. My feeling tells me, that we will need to
handle them more like the other tabs (they should be persistent,
it should be possible to open multiple 'New game' tabs, etc.).