Fixes #802 - Author of a game gets automatically directed to a new game
When an author of a currently active game was going to the root URL, the
automatic table join directed him to create a new game despite the WS
table status pointing to the right game. The client table plugin looked
at whether he was the next author before looking at the fact that he was
not in the active game - so the client believed it was his turn to
create a new game, rather than understanding that he had already created
the game and just needed to rejoin it.