Make the page not reload when switching the tabs (closes #810).
The game also doesn't reload after creating a new game. The URL stays
the same all the time, which isn't ideal, but there are ways around
that and we will probably want to do something about it in a separate
task.

The $.cardstories.reload method has been modified so that it doesn't
reload the page anymore (I decided to not change its name even though
it doesn't really reload the page anymore - as it still loads another
game, but I'd be happy if somebody suggested a better name).

The bootstrap phase has been split into two parts - one that only needs
to happen the first time the page is loaded ($.cardstories.bootstrap),
and another one that has to happen each time a game is loaded
($.cardstories.load_game). The $.cardstories.reload function is just
a wrapper around $.cardstories.load_game that takes care of reconstructing
a clean state of the DOM.

The JS plugin API had to be adjusted similarly. The init function of the
plugins is only called the first time the page is loaded, but the load_game
function is called every time a new game is loaded.

I'm not very happy with the names and I'd welcome any suggestions. Also, it
might be good to settle on a single pair of names (bootstrap/load_game vs
init/load_game).

In the bootstrap function, a HTML copy of the contents of the root is saved
onto the root itself and is used later when loading a new game when the
contents of the root container are emptied and reconstructed from this saved
HTML copy.

Looking at the diffs you might see what seems like a large modification of
cardstories.html, but I actually only added a wrapper div around the structure,
since I feel it is a bit cleaner when we are saving and restoring it when
loading games.

In order to support tables, several of the create_* functions take a new
parameter 'previous_game_id'. I couldn't come up with another way of doing
this that wouldn't be a hack, and this supports the idea that tables should
be part of the core code.

As a minor change, I removed the .cardstories_tab CSS class from the + button
(.cardstories_new), since it was more straightforward to count the number of
opened tabs in tests this way and it didn't seem to break the layout.

While testing the new reload method in Chrome 17 on OS X, I sometimes get
a weird result where the deck (top left part of the board) becomes invisible.
I suspect this might be a bug in Chrome, since I cannot find out what could
make the element invisible through the Chrome console, and Chrome just seems
sort of buggy on OS X. I didn't notice this problem in Firefox or Safari.